+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Unreleased
+
+
+### 2.1.0 2018-07-31
  - Adding scripts for managing response on quarantine queue
  - Remove unchecked folder and fixed log statements for AV scan